font_size presentation parameter is float, not int

+++ lib/lcg/parse.py
@@ -135,7 +135,7 @@ class Parser(object):
     _PARAMETERS = (('parameter', 'page_header', 'HEADER', None,),
                    ('parameter', 'first_page_header', 'FIRST_PAGE_HEADER', None,),
                    ('parameter', 'page_footer', 'FOOTER', None,),
-                   ('presentation', 'font_size', 'FONT_SIZE', int),
+                   ('presentation', 'font_size', 'FONT_SIZE', float),
                    ('presentation', 'font_family', 'FONT_FAMILY', None),
                    ('presentation', 'noindent', 'NOINDENT', bool),
                    )
